requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
Django==3.1.2
imageio==2.9.0
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
plotly==4.13.0
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
retrying==1.3.3
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tifffile==2020.12.8
toml==0.10.1
wrapt==1.12.1
